pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spreadsheetbot"
version = "2.0.1"
authors = [
  { name="Egor Dubrovin", email="dubrovin.en@ya.ru" },
]
description = "Google Spreadsheet-based Telegram Bot Package"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  'python-telegram-bot>=20.0b0',
  'pandas',
  'gspread',
  'gspread-asyncio',
  'aiohttp',
  'filetype'
]

[project.urls]
"Homepage" = "https://github.com/twobrowin-study/spreadsheetbot-lib"
"Bug Tracker" = "https://github.com/twobrowin-study/spreadsheetbot-lib/issues"